// project/src/types/filters-type.ts
import { FilterTypeItem } from './';

type FiltersType = {
  Photocamera: FilterTypeItem;
  Videocamera: FilterTypeItem;
  Digital: FilterTypeItem;
  Film: FilterTypeItem;
  Snapshot: FilterTypeItem;
  Collection: FilterTypeItem;
  Zero: FilterTypeItem;
  NonProfessional: FilterTypeItem;
  Professional: FilterTypeItem;
}

export default FiltersType;